backend/payment/requirements.txt
aioredis==2.0.1
anyio==3.6.1
astroid==2.12.10
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
cleo==1.0.0a4
click==8.1.3
crashtest==0.3.1
Deprecated==1.2.13
dill==0.3.5.1
fastapi==0.85.0
h11==0.14.0
hiredis==2.0.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pptree==3.1
py==1.11.0
pydantic==1.10.2
pylev==1.4.0
pylint==2.15.3
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
python-ulid==1.1.0
PyYAML==6.0
redis==4.3.4
redis-om==0.0.27
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
tomlkit==0.11.5
types-redis==4.3.21.1
types-six==1.16.21
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
wrapt==1.14.1
